Fix example_startup_trace build on windows
trace_data[0] is technically undefined behavior if trace_data is empty
(which happens on windows because IPC do not work, but can also happen
on linux if the system tracing service has not been started).

unistd.h does not exist on windows and is unused on linux.

Startup tracing is not supported on windows, because the system backend
does not work there, but this at least fixes the build.

Change-Id: I55cab3832c05a90c6baaceec301ed4b4dcb030cc