tp: enable write() syscall tracking, truncating durations when combined with atrace slices
Atrace slices are emitted as begin/end events written into the tracefs
trace_marker. If we're tracing syscalls, the reconstructed atrace slice
would start and end in the middle of different sys_write slices (on the
same track). Since trace_processor enforces strict slice nesting, we
need to resolve this conflict. The chosen approach is to distort the
data, and pretend that the write syscall ended at the atrace slice's
boundary.

In other words, this true structure:
[write...].....[write...]
....[atrace_slice..].....

Is turned into:
[wr][atrace_slice..]
...............[wri]

I couldn't think of a better alternative that preserves the true
sys_write durations, short of kicking those slices off to a different
track, but such mutation isn't readily supported by trace_processor, and
would likely be a maintenance burden.

Open to suggestions since I don't like distorting the data.

Change-Id: I30544f73f3d4aacb307a3be5b8e1375d8d89758d