tp: sched: fix preempted and idle state parsing (by reimplementing the parser)
In 4.14, upstream changed sched_switch to be more similar to
/proc/pid/status printing. So there is only a specific set of bits that
is reported (TASK_REPORT macro) instead of the full set of state bits
defined in the scheduler header. This also means that the sched_switch
preemption flag that used to be TASK_STATE_MAX (one past the highest
state bitmask bit in the scheduler implementation) is now
TASK_REPORT_MAX (one past the highest state that procfs reports):
* https://github.com/torvalds/linux/commit/efb40f588b4370ffaeffafbd50f6f
* https://github.com/torvalds/linux/commit/3f5fe9fef5b2da06b6319fab81230

Perfetto tried to handle this, but was incorrect in handling of
TASK_REPORT_IDLE and TASK_REPORT_MAX (i.e. preemption). Further, there
was additional confusion between "dead task" and "idle task" since they
have the same flag bit (but the earlier is masked off in 4.14+ as it's
not part of TASK_REPORT).

I've reimplemented the class to further disassociate the parsed
representation from the raw bitmasks, so all of the kernel version
specific assumptions are now in the FromRawPrevState factory method.

Known issues remaining as of this patch:
* pb -> systrace conversion isn't correct unless the trace is from a 4.4
  device. This is because trace_processor destorys the kernel version
  knowledge before traceconv starts calling into conversion functions.
* UI looks a little bit weird (check out kthread tracks) since the
  "Idle" text is still rendered in black.

Bug: 247222275
Change-Id: I7d1bbc252dbb9d60fce60bd73597d3ca0e821f5d