Support using system buildtools in generating the SDK
This change adds option --system_buildtools to tools/gen_amalgamated
that can be used to make the script use buildtools preinstalled in the
system, other than the ones installed with tools/install-build-deps.
This is option used to generated the SDK source inside the ChromeOS
build environment, where downloads are disallowed during build time.

Bug: b/247942837
Change-Id: I877112ad941d16f186e8024a838d6a7b3ead2619