Add a flag to call data_source callbacks out of lock
To avoid lock order inversion issues, we try not to call into embedder
code under a lock, where possible (because we can't prevent the
embedder from doing the opposite: holding a lock and calling into
Perfetto code).

This CL adds a flag for a data source to allow calling callbacks
without a lock. It also allows unlocked callbacks for the TrackEvent
data source (which is responsible for most deadlock issues in Chrome).

Bug: b/245121030
Change-Id: I768ef738d38ed6003ae217d9b5f6cf9408791439