Support passing tags to queries.
Allow users of engine.query pass an optional tag (usually a component
name) and plumb this tag to trace processor and emit it in the metatrace
RPC_QUERY event. This allows post-factum analysis of trace processor
load, including grouping all of the queries by tag.

Introduce EngineProxy class to make it easier to pass an EngineProxy to
an component and have all queries be automatically annotated.

R=hjd@google.com

Change-Id: I948bbfc3d463f2e38c74f98d778e930d55303572