Replace substr with substring in router.ts
The former has been deprecated; this CL replaces the calls to substr
with calls to substring. The difference between the two is that the
second parameter is the last index instead of substring length, but the
only two-parameter call has 0 as the start, so the method name can be
changed without extra fixes.

Change-Id: Iaad3fb4d95fbcf43a7711be444419043c88fd113