tp: Extract return value of syscalls
The return value is recorded for sys_exit events but never used when
processing the trace. Make sure to use it when available so trace
analysis can also be based on returns of syscalls.

This requires to split the function previously shared
between parsing sys_enter and sys_exit events as we need
to access the actual protos to get the ret field.

Change-Id: I452b5723c6bfd757e093de81c512eb1676b3b94c