tracing: Send triggers on connection
The newly introduced ActivateTriggers only sends triggers to the
currently connected backends.

Since perfetto::Tracing::Initialize() is not blocking, it can happen
that a backend takes a while to connect.

If a registered backend is not connected when ActivateTriggers is
called, the system will remember that and send the trigger when the
backend connects.

An expiration must be provided to avoid sending triggers if the tracing
service takes too much to connect.

Bug: 259929612
Change-Id: I183ef3ae04e960797adc835e137f854d0ebd409d
(cherry picked from commit 037282790b01e71cf004c3bc05b035d62eff8206)
Merged-In: I183ef3ae04e960797adc835e137f854d0ebd409d