ftrace: Collect new fds
Keep any newly created filedescriptors as sets
in metadata whenever ftrace sys_exit events are enabled.

This is in preparation of a follow-up patch which will add
scraping of filedescriptors to ProcessStatsDataSource so tp can
do further augmentation of syscalls as part of its postprocessing
and provide context on file descriptors used by a process.

Change-Id: Ie41964b4fd48bf2fb73b6dfd1cfe8753506ab7be