JS-side higher-level aggregation for pivot table
When several pivots are selected, pivot table used to generate a
complicated query in order to compute aggregate values for all levels
of nesting. E.g., when pivoting on category and name, aggregate values
would be computed over (category) and (category, name) windows.

This CL changes the implementation so that only fine-grained
aggregations are computed on SQL level, and the values for the rest
are computed on the JS side. This should simplify changing the query
generator further, for example, to add support for "average"
aggregation.

Bug: b/260561224
Change-Id: I422cca098ff6785ca06f84ad0555cb354e697c29