Pivot table: remove preaggregation
I have investigated whether the preaggregation is necessary in order
to achieve reasonable performance, and it does not seem to be the
case: version of query without preaggregation performs about the same
way on a random sample of 300 traces.

Therefore, it's useful to get rid of preaggregation in order to
simplify query generation for further feature development and
maintenance. This CL is doing exactly that.

Bug: b/259545074
Change-Id: I34f502da33f10685e10778ee37a871824264dbb1