ps: Optimize mapping of tid to pid
Right now data is duplicated between maintaining a mapping
of tids to pids (`tids_to_pids_`) and also a set of pids
which have been seen (`seen_pids_`).

This can be deduplicated by swapping the type stored in
`seen_pids` with a struct holding both the pid and parent
but only considering the pid for equality and insertion
order. By use of an implicit constructor, use of seen_
remains identical to before but if needed a find operation
can be used to not only check if a PID is known but also
what parent it belongs to.

Change-Id: Ia9c1dd29b490a3c0a797e0631efccd3712488bf0