tp: fix SQL amalgamation deps withou depfile
depfile was a fragile and hacky way to do dependency tracking and broke
docs generation as the docs target was not taking into account the
depfile. Instead, switch the SQL source sets to be actions (similar to
protos) which means GN does proper dependency tracking on them.

There is no diff for either Bazel or Android.bp as a result of this
change as they were already correctly handling dep tracking.

Change-Id: I52f4c99d75108bf5f788922101b239c96d1bffe0