tools: overhaul and simplify transitive dep propogation
Before now, we had transitive dep propogation for some targets
(source_set, proto_library) but in a way which made it very hard
to follow. Also because it was only for some target types, it is
difficult to reuse.

Instead switch to a model where we propogate deps transitively unless
the target type specifically does not allow it. Add some helper
functions to take care of filtering the transitive list and make use of
them in the generators.

Change-Id: Ic24b7ea1ac172adb4c1349409f3e23cfc669be05