ftrace: Add atrace aware rules in ftrace/print filter
Today the filter rules just allow for a prefix to be specified. This is
not expressive enough if we only want to filter specific atrace counters
by name, because the pid is not predictable:

```
C|<pid>|interesting_counter
C|<pid>|not_interesting_counter
```

This CL implements a way to match the string after the pid with a
prefix.

Example:

```
ftrace_config {
  ...
  print_filter {
    rules {
      atrace_msg {
        type: "C"
        prefix: "interesting_"
      }
      allow: true
    }
    rules {
      atrace_msg {
        type: "C"
        prefix: "not_interesting_"
      }
      allow: false
    }
  }
  ...
}
```

Change-Id: I842af43dc2e43ec3b13883ebdd13d26ba2d4f9c8