ftrace: traced_probes should not use tracing_on for tracefs is in use
tracing_on is a trigger for enabling ftrace ring buffer and should not
be used as an indicator whether tracing is using by something else.
The CL replaces the check, treats tracing as available if current_tracer is "nop". Also the CL enable tracing_on if it
is off before activate ftrace config, and restore it to the original
state at the end.

BUG:261149730
Change-Id: I24b8f53d727e819f0947495325896999867fea95