ui: Remove the controller dispatch() layer
The existence of two different globals.ts files:
- ui/src/controller/globals.ts
- ui/src/frontent/globals.ts
is a source of confusion and, now the controller worker is no more,
completely unnecessary. We should remove ui/src/controller/globals.ts.

The first step for that is to remove the legacy dispatch() method
on controllers/globals and instead have controllers invoke the
frontent/globals version directly.

Change-Id: Iea6083f2834d4284b63c17d40b3a021892eaf707