tracing: Fix race condition reported by TSAN
Most of the fields of `DataSourceState` are
written on `SetupDataSourceImpl`, before doing a release store on
`valid_instances`, so they can be accessed just after an acquire load on
`valid_instances`.

`trace_lambda_enabled`, on the other hand, is set on
`StartDataSourceImpl` (i.e. after the release store on
`valid_instances`), so TSAN thinks that there's a data race.

Let's mark it as atomic.

This should not have any change in behavior (relaxed), it's just for
documentation purposes and to make TSAN happy.

Change-Id: I828ba7901a2facf0d04af59762e7066571df28e6