Add a thread_executing_span stdlib
A 'thread_executing_span' is thread_state span starting with a runnable slice
until the next runnable slice that's woken up by a process (as opposed
to an interrupt).

Note that, some scheduled work in a thread doesn't have a distinct
runnable slice, we just seamlessly transition from running the previous
task to the new one. With the newly added spurious_sched_wakeup table,
we should be able to detect this, but for now, they are simply ignored.

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*TablesSched:.*'
Bug: 231758687
Change-Id: Ib28f2674c463c06f38a74a9a4f174fc3b4d11ca2