tracing: Allow initializing tracing backends separately.
With this commit, users of the SDK can call Tracing::Initialize()
with different backends (system or in-process) separately.

Also, the system consumer backend can be initialized in
Tracing::NewTrace, only if we're sure that we need that.

This is going to be especially useful for the shared library, where it's
easy to imagine having different code bases calling the initialization
independently.

Change-Id: Ic36414b7186b9e0b8569ffd5a53bdcb3e79ff560