perfetto: switch to being a C++17 only project
Now that it's Feb 23, and we've given plenty of warning to people both
internally an externally, it's time to close out C++11 support and start
allowing C++17 features into the project.

This CL removes the last pieces of C++11 support from the build
files. In follow-ups, we can remove all the PERFETTO_IS_AT_LEAST_CPP17
usages and cleanup the codebase.

Bug: 257064316
Change-Id: I19ae1c452c8ef31a50bf7d1ee4105c3b783c5e4b