Trace processor support for entity residency packets.
We just pull the total time in state to start with as probably the most interesting.

We put all the counter tracks in a group in the UI since they're quite numerous. We add a prefix to each track to let us recognise them but then remove it again in the UI code to reduce clutter.

Screenshot: https://screenshot.googleplex.com/A5kYxwZqThAacyT

Bug: 269589187
Change-Id: Ib202f2afc6897c5a69bd0d0f4e7fc28bd64b174e