cts: attempt to deflake TracedPerfCtsTest
Instead of relying on traced_probes happening to run at the time of
periodic sampling, sample all context switches and let the profiler
filter that down to traced_probes.

Slightly concerned about pressure on kernel perf buffers on slow
emulators and whatnot, but imo the pragmatic approach here is to submit
and observe.

Bug: 270372997
Change-Id: I814d70e9de4150a3109987fbfbda22e0d9eafd44