tp: lift limitation of 4GB chunk count in BumpAllocator
Seems like for large non-proto traces, because we do not have
compression, the total memory managed by the BumpAllocator can exceed
4GB.

Because of this, make all the allocator ids 60 bits instead of 32. This
ends up having to allocate 8 extra bytes per
TracePacketData/TrackEventData for the packet size but to avoid
significantly increasing the complexity, it's fine for now. If we see
these 8 extra bytes are a big problem, we can revisit this.

Bug: https://github.com/google/perfetto/issues/456
Change-Id: Id75a0aa331e1ca60e73530efa50e75743f780945