trace_sorter: Fix build on windows
bit-fields layout is implementation defined. MSVC (and clang-cl on
windows) behave differently than GCC and clang on linux: when they
encounter a field with a different type, they don't "merge" it with the
previous one.

This created a problem for TimestampedEvent, because the size wouldn't
match a static assertion.

Fix the problem by using the same type for all the bit fields in the
struct.

Change-Id: I9017332b166557e7a4da9bd3ffb5d39a92eec9f4