Introduce Icon component to abstract icons
Recent change aosp/2470800 broke a couple of things where I relied on
variations of icons to be visually distinguishable (e.g. 'star' and
'star_outline'). Abstracting a component for the icon seems to be a
reasonable way to mitigate those problems in the future; this CL
introduces such component and uses it in query history.

Change-Id: I0c0d401c1274fa9596da7cc5dffcc9174b10fcdd