shared_lib: Implement ResetForTesting()
Problem: we'd like to support tests for the shared library APIs in the
same process with other unit tests. But the shared library relies on
global state which cannot be uninitialized.

The C++ tracing SDK has a similar problem and it's solved with
Tracing::ResetForTesting(). Implement a similar solution for the shared
library.

ResetForTesting is not exposed as an ABI from the shared library, it's
only exposed with internal headers (which are going to be statically
linked by tests in future CLs).

Bug: 184929776
Change-Id: I97e72ea0eddf065e678432d985b374ee119142e1