ui: Add schema reflection for EventSet
For EventSet we need to inspect the schema of a trace_processor
instance to know the tables, columns, foreign-keys, parent/child
table relations etc. This adds the initial bare bones version of this.

At the moment the only useful behaviour is to get the columns for
a couple of manually specified tables. The eventual idea is that
this would be responsible for:
1. retrieving the whole schema
2. automatically figuring out which tables to join
   (or which C++ views to use) to get a given set of EventSet keys

The implementation is split into some POJOs and a class
(DatabaseSchema) with the actual behaviour both to make it
possible to test the behaviour without a live engine as well as to
provide a good split point in case we ever want trace_processor
itself to provide the schema.

See go/perfetto-ui-eventset-design and the prototype at aosp/2320374
for additional context.

Change-Id: I06a69352d22a6c9d08b43fb9485e337b2564f7a7