[metrics] Improve android_binder query performance
The sub-query to detect broken binder txns was killing performance.
Some field traces were taking up to 10s. With this change the same traces
complete in <200ms. There was also a large local trace of ~800MB that was
impossible to run this query on, taking >30mins
(never actually waited to completion). Now that completes in ~15s.

Test: tools/diff_test_trace_processor.py \ out/android/trace_processor_shell \
--name-filter '.*binder.*'

Change-Id: I56667dc77f17c06f9653dc204c79ab08078690a7