Anti-flickering logic for bottom tabs
The current implementation of the bottom tabs has the side-effect
of unpleasant flickering due to a combination of two factors:
- Adding tabs to the tab list is instanteneous
- Fetching the necessary data for them (which usually involves TP
  queries) is not.

This leads to a lot of quick visual noise, when the newly-added tab
transitions from showing "loading" to showing the actual data.

This patch implements BottomTab.isLoading API and a concept of "pending
tab list": all newly-added tabs go into a queue of pending tabs. These
tabs will be moved to the actual list of displayed tabs:
- in order which they were originally added,
- and when either:
  - The tab finishes loading.
  - 50 ms elapsed since the tab was added

50 ms was selected as half of the guideline 100ms target for responsiveness,
leaving another half for the rendering pipeline to fit the overall
interaction latency under the target.

Bug: b/273215738
Change-Id: I640bed2d47b5f596200432e1423d0257b059c215