traced_probes: Refactor statsd_client
A lot of the over general code in statsd_data_source
was based on the idea that we would eventually add a second
`StatsdBackend` which, rather than execing `cmd stats data-subscribe`,
used the related binder API[1]. Since that API was also based around
file descriptions a lot of the code could be shared.

As it happens that didn't work out and the new binder based API
we are actually going to used is callback based so the
generalisation was unnecessary. Instead the callback based API
can have a DataSource subclass all of it's own.

This CL:
- collapses ExecStatsdBackend into the main class
- extracts the code we'll actually want to share into a
  'common' file
- Does a few other random cleanups (removing the unused
  GetWeakPtr static method).

[1]: https://cs.android.com/android/platform/superproject/+/master:frameworks/native/libs/binder/ndk/include_cpp/android/binder_interface_utils.h;l=243;drc=affab527c820c06d3a6b3c4451aaaa744e652eb7

Bug: 268661096
Change-Id: If6701b1f677028573495ccc0cab3e67ce615b15e