Factor out event writing code into a separate function
Template instantiations of TraceForCategoryImpl often can't be folded
by ICF because their binary code differs a little bit. The difference
is due to trace event args that are different in almost every
callsite.

This CL factors out the event writing code that doesn't depend on Args
into a separate function that can easily be shared between many
trace macro invocations. The argument writing code that is left in
TraceForCategoryImpl is smaller and so the total overhead is smaller.
This CL saves ~90KB of binary size in Chrome on Android.

There is an effect on runtime tracing overhead because of the extra
function call, but it is pretty small:

Benchmark                                   Before           After
------------------------------------------------------------------
BM_TracingTrackEventBasic                   155 ns          162 ns
BM_TracingTrackEventDebugAnnotations        230 ns          237 ns
BM_TracingTrackEventDisabled               1.36 ns         1.36 ns
BM_TracingTrackEventLambda                  195 ns          202 ns

Bug: 268328757
Change-Id: I7c848339dbaa6f96522f53c130a4633b543931d3