Tweak Select All/None button text in mutliselect widget for clarity.
When search is active (search/filter box is not empty), changed:
- Select All -> Select Filtered
- Select None -> Clear Filtered
Also changed the search box placeholder from "Search..." to "Filter
options..." to match the "Filtered" terminology on the buttons.
The reason for going for "Filter options..." rather than just
"Filter..." is because the multiselect is used for filtering in the
ftrace panel, so it makes sense to specify what we're filtering here.

Change-Id: Iaee1aa35a71328225cf9ea1a1a895b6c0396b558