[metrics] Add monitor contention thread state information
Similar to the android_binder metric, the android_monitor_contention
slices are enriched with thread state information to further
investigate long contention slices.

One slight quirk is that the thread_state breakdown is on the
'blocking' thread as opposed to the 'blocked' thread, i.e. for
each contention slice (or node as the metric is called), we add
a breakdown of the thread state of the thread holding the lock
and hence blocking the slice. This is more useful in analysis
as the 'blocked' thread is almost entirely sleeping and this
information is uninteresting, whereaas, what the 'blocking'
thread was doing yields interesting insight.

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*monitor.*'

Change-Id: Ie5c6e0d9908ff7c02aab19497a04dde5f501e303