protozero: Add a tool for bytecode filter inspection
Sometimes it can be useful to inspect which proto fields are allowed by
a precompiled bytecode filter. Previously the only easy way to do that
would be look at the proto schema which was originally used to generate
the bytecode. However, sometimes the original proto schema isn't
available, which makes it hard to verify what a given bytecode filter
does.

This patch adds an option to `proto_filter` to take a precompiled
bytecode filter and apply it to a schema. The resulting output shows
which fields of the schema are allowed by the bytecode.

Example: proto_filter -s protos/perfetto/trace/trace.proto -f /tmp/bytecode

Bug: 259328844

Change-Id: Ie088e79ff9abc20dbf4a2277b9240eaecc398cfa