heapprofd: Fix problem with continuous dump
An heapprofd client serializes all allocations/frees and assigns a
sequence number to them.

The allocations/frees are sent to the UnwindingWorker. To improve
performance, the frees are batched.

The UnwindingWorker sends them to the HeapTracker, which processes them
in the correct order (based on the sequence number) and produces
callstack dumps.

In normal operations (when continous dump mode is disabled):
* The data source is stopped.
* The batch of frees in UnwindingWorker is drained.
* HeapTracker receives all operations and sorts them.
* A single dump is produced.

With continuous dump mode, dumps are produced without draining the batch
of frees in UnwindingWorker. This means that HeapTracker is not able to
process new allocations if there has been a free (because it sees a
missing sequence number).

Fix the problem by draining the batch of frees in UnwindingWorker before
emitting a continuous dump.

Reported-at: https://github.com/google/perfetto/issues/477
Change-Id: I8fb8381a476406fa554ae37d067d01ed1ae5e379