Add view for Android BatteryStats events/counters
The counters are used for things that can only have one state at a time
(like screen is on or off). This turns the counters into spans with
human readable names.

The events can be used for things that happen once, or things with a
start and stop interval that either doesn't map well to the counters or
can have multiple overlapping (e.g. jobs). This converts such events
into spans.

Test: ui/run-dev-server and query metric
Change-Id: Ib600b9426482a22e6c1a393b983e17f9eac257d6