protozero: Avoid DCHECK when parsing malformed packet repeated
The protozero decoder treats packed and non-packed repeated fields
differently.

When a decoder parses a packed repeated field, it expects a
kLengthDelimited wire_type, so it calls Field::data(). Field::data() has
a DCHECK if wire_type is different. On optimized builds everything is
fine (although kind of fragile): the DCHECK is ignored, the
PackedRepeatedFieldIterator sees a 0 size and behaves as if the field
was empty. On debug builds, perfetto will crash on the DCHECK.

This commit explicitly checks the wire_type before calling Field::data()
because:
* It avoids a DHECK on debug builds.
* It avoids relying on the fact that Field::size() is 0 for non
  kLengthDelimited fields, which is kind of subtle.

It also adds a unittest (that would hit the DCHECK with the old code).

Change-Id: I71821c13e423a17aa6105ebe393dd2579793dc6e