tp: rewrite integration of tables with SQLite
This is a major refactoring of how we register tables with SQLite. We
now proxy all registrations through SqliteEngine which allows us to
keep track of what we've registered with SQLite in a follow-up CL.

Moreover, this CL also overhauls how we create the sqlite3_module struct to make it clearer to follow what is going on.

This is a noop refactor and does not change the behaviour observed by
anyone using the SQL API.

Change-Id: I7384d4010c535881fd00e1264b31288bbaddf10f
Bug: 261195778