Add a template metric for network activity query.
It is often quite useful to see networking aggregated into regions that
indicate whether the interface should be on or off. In the simplest
case, it can be an easier way to view large quantities of instant
network events. In other cases, it can be used to proxy the network for
power or similar analysis.

Test: diff test and ui-dev-server
Change-Id: I0c58b59d082fe8de640f32fa2a65483e9e9a9e0c