Add metric tracks for useful battery stats states
This adds metric tracks for {cell, wifi} {status, radio, strength}. The
counter forms of these are not easy to understand alone (e.g. remember
that data_conn=13 is LTE).

Note: the underlying view is updated to partition on the track name,
allowing the WHERE condition filter to be pushed down. This reduces the
query time on some large traces from >100ms to 25ms.

Test: ui-dev-server
Change-Id: Ia6a556726afe7fba4fe9e2b5e4d7793c468ad1e9