[metrics] Fix android_monitor_contention metric
The thread state (and blocked functions) breakdown of the metric was
broken because of overlapping thread state span joins on the
'blocking thread' (thread holding the lock).

In a contention chain where Thread A is blocked on B, the metric we
collect is the thread states on B during the span of A, as opposed
to the thread_states on A itself. This is so that we can identify what
the lock owner is doing while the waiter is blocked. Unfortunately,
multiple waiters on the same lock caused overlapping span joins which
led to undefined behavior.

To fix this, we only collect thread state information on nodes with
waiter_count = 0. This ensures that we have unique time spans on the
blocking thread.

Also fixed s/LIKE/GLOB/g

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*monitor_contention.*'

Change-Id: Ic9abd0fbb139cce7766cbd72356552c9280a3885