tp: introduce tokenizer for SQLite statements
This CL adds an "SqliteTokenizer" class which allows tokenizing SQLite
SQL statements into tokens. The main function is just a minified copy
of the |sqlite3GetToken| function (i.e. with keyword specific parsing
removed to reduce the numbeer of constants copied).

The class layered on top of this function allows for C++ friendly use.

Change-Id: Icc10bf0aea60b16c013a38c8afa29bf4d41ebb64