UnixSocket: support the vsock address family
This change adds the support of vsock family to UnixSocket. A vsock
socket name is in the format of vsock://$CID:$PORT.

For example:
`PERFETTO_PRODUCER_SOCK_NAME=vsock://-1:10000 traced` starts traced with
the producer socket listening on the "ANY" CID and port 10000.
See vsock(7) for special CID values that may be used for vsock.

Bug: 284258446
Change-Id: Ie6f8e4a9f55c0550dba48818e6e26642763fad8d