Make trace smaller in CtsPerfettoReporterTestCases
The reporter test only works with a trace that's smaller than 1Kb.
Recent additions to the tracing service (chunk usage stats) caused the
final trace size to be bigger than that on some devices or builds (the
device and build names are included in the trace).

This commit fixes the problem by disabling a bunch of tracing service
feature. This reduces the tracing size on redfin from 1020 bytes to 400
bytes.

Bug: 282508742
Change-Id: I6c17948e7983828eb4f11bf760f71d8055b777ba