Use bigints directly in time formatting functions
In the UI there exists a trio of time formatting functions, including:
- formatTimestamp
- timeToCode
- timeToString
Previously, their input was a number, but since switching to bigints,
these were patched to take bigint time but they simply converted this bigint to a number and processed it as before.
This CL fixes this patch so no conversion to a number is necessary.
This fixes an issue where large timestamps lose accuracy when printed
with either formatTimestamp() or timeToCode(). timeToString() is the
exception as it purposely loses precision for displaying durations
mainly in a more human readable format.

Change-Id: I18ede3744609c8316a80c137028f6f53395b06b1