DebugTrackV2: links to slices.
For the cases when the debug track is generated from the slice or
thread_state table, generate a link go to that slice / thread_state
in the trace.

The implementation looks at the presence of a few key columns (id,
ts, dur, track_id for slice; id, ts, dur, utid for thread_state),
queries these tables, validates the consistency of the data and
shows the link when the data is indeed consistent.

Bug: b/273215738
Change-Id: I50c273dfe68bb99ea3426179e0b979b3e0954beb