Recursive calls unrolling
When running trace_processor in the UI, max recursion depth of the
Perfetto SQL functions is limited to just under ~200 (the V8 stack
is limited to ~1k frames and due to a combination of SQLite and Perfetto
calls take ~7 frames per recursive Perfetto SQL call).

To mitigate this, this patch implement recursive call unrolling.
When a recursive call is detected, we will unroll the recursion using
the following two-pass approach:
- First pass will add the arguments into a queue to be evaluated. During
  the first pass, recursive calls will immediately return and its
  arguments will be added to the queue.
- Second pass will evaluate the statement according to their order in
  the stack.

Change-Id: I94ee859221d9dd11ee8de822f5b0c6f9b7a6dc5d