Add new track for top level scroll jank summary.
This track includes slices for
* Scrolling intervals (total time scrolling regardless of the
number of scrolls
* Jank duration during the scrolling interval
* Specific causes of scroll jank

See screenshot: https://screenshot.googleplex.com/8vtrtwa6rqRWcUa.png

Additionally migrate to a unified Scroll Jank selection type; we can
migrate further to a single Custom Track selection type in a future
patch for further reuse. In the same vein, added a generic details
panel for scroll jank, that we can migrate out into a generic
Custom Details Panel in a future patch.

NOTE: The event latencies track will be revised in this patch:
https://android-review.git.corp.google.com/c/platform/external/perfetto/+/2576994.
Changes may include where the track is rendered in the UI.

Bug: b/225885015
Change-Id: I220b10a12621000a1cc1307d12721a51137379a7