ipc: infer socket family from address
Up until now the socket family used by the IPC layer was
hardcoded depending on the plaform (AF_UNIX on Linux/Mac/Android,
TCP on Windows).
Upcoming changes for VMs require more flexibility.
This change make it so that the IPC layer (and hence the env vars
PERFETTO_*_SOCK_NAME) figures out the family from the address.
This has also the benefit that we can now forward the consumer socket
over TCP on a Linux system.

Test: PERFETTO_CONSUMER_SOCK_NAME=127.0.0.1:1234 traced
  PERFETTO_CONSUMER_SOCK_NAME=127.0.0.1:1234 perfetto -c :test -o /dev/null
Bug: 284258446
Change-Id: I4fc9150cc0fe4437dbf2d563be7b969964c3e819