traced: allow cross-session cloning for bugreports
Prior changed introduced generalized support for CloneSession
and rewrote the --save-for-bugreport feature on top of that.
Unfortunately we didn't notice that the CloneSession feature
is very strict on UID checking, enforcing that source and target
consumer UID match. This is good in general but not desirable
(And a regression) for the case of bugreports, where the two
uids differ.
In this case bugreport_score acts as an "I am fine exporting
this trace to other uids".

Bug: 286908237
Bug: 260112703
(cherry picked from https://android-review.googlesource.com/q/commit:5457722190e8cad7dddbedc71b8957ab45402a6c)
Merged-In: I99bd22aef0bfa402cb24559b8515c5bdf0b48ba9
Change-Id: I99bd22aef0bfa402cb24559b8515c5bdf0b48ba9