[stdlib]: Add a table with flattened slices
Flattened take an ancestor slice partition and projects all the
descendants to the root slice. The resulting table can be span
joined with other tables like 'sched' to calculate self time of
slices.

Note that we already have a custom CPP table, 'experimental_flat_slice'
that provides practically the same result. As a follow up, that will be
deprecated in favor of this since this is more flexible and scalable.

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*slice_flattened.*'
Change-Id: Id63fbc407fbc8bc9fff12f7bc102eeccfcfd8381