Add --stdlib-sources argument to check_sql_modules
We need to be able to run tools/check_sql_modules.py on the trace processor chrome stdlib, which will now live in the Chromium repo. In order to run the script from Chromium, we need to be able to provide the path to the chrome stdlib. For example, tools/check_sql_modules.py --stdlib-sources base/tracing/stdlib

Bug: b/287233783
Change-Id: I25faa3124d3579bed11b5fd5c7772843ea60d2e4