traced_relay: implement the relay service
This CL implements the relay service as a new binary, traced_relay, for
taking the place of traced in a guest VM and relaying the producer
connections to the host tracing service. The relay service listen to the
producer socket and connects to the "relay socket" to the host. Socket
connections are opened in pairs. Trace data and control commands are
relayed back and forth between the connection pair.

        tracing service
               ▲    PERFETTO_PRODUCER_SOCK_NAME
 relay socket  │
               ▼   PERFETTO_RELAY_SOCK_NAME (tcp or vsock)
        relay service
               ▲   PERFETTO_PRODUCER_SOCK_NAME
producer socket│
               ▼   PERFETTO_PRODUCER_SOCK_NAME
          producers

The relay socket is used only to connect to the host producer port. The
consumer port isn't supported. Peer credentials (PID and UID) is sent in
the wire_protocol proto by the relay service.

Setup example:
* Start traced with producer endpoint on localhost:10001:
  `PERFETTO_PRODUCER_SOCK_NAME=127.0.0.1:10001 traced`
* Start the relay service that uses unix socket for local producers and
  localhost:10001 for relaying to the remote tracing service:
  `PERFETTO_RELAY_SOCK_NAME=127.0.0.1:10001 traced_relay`
* Start traced_probes that connects to the unix socket to the relay
  service:
  `traced_probes`
* Collect a trace using the default consumer socket.

The trace data provided by traced_probes will go over the unix socket to
the relay service, forwarded to localhost:10001 and then get to the
tracing service.

Bug: 284258446
Test: Unit tests in a followup CL

Change-Id: I3e13b5d69b0cb78b417ccd3cc1523901b2ebebdc