Fix unused return value
The local symbolizer now needs to be built in chromium as well in order to run diff tests for the chrome stdlib. Calls to symlink() in the local_symbolizer_unittest previously ignore the return value, which fails to compile in Chrome due to the -Wunused-result flag. We now use the return value with an expectation.

Bug: b/283962174
Change-Id: Ieed02194b29cfbc343a9924de1eabdff033fcb7a