Remove additional join from android_blocking_calls_cuj_calls
This join was filtering slices to be only in cuj processes. This was meant to be just an optimization to have less slices in the next table operations, but it ended up being very expensive.

On a test trace full of binder transactions, the metric generation time went from 399s to 13s.

Bug: 286095857
Bug: 290780193
Test: Test:  tools/diff_test_trace_processor.py out/linux_clang_debug/trace_processor_shell --name-filter Android:android_blocking_calls_cuj --keep-input
Change-Id: I2b28d3e2b47c2ab1f178bb25aad109fdd41f7fb0