shared_lib: New macro to initialize struct PerfettoProducerInitArgs
Currently PerfettoProducerInitArgs has only one field, so it can be
initialized with `{0}`.

In the future, if we add more fields to PerfettoProducerInitArgs (we
can do that, it's not part of the ABI), a `{0}` initializer will cause
warnings.

Let's add a macro to provide an initializer. The macro can be changed in
case new fields are introduced.

Bug: 184929776
Change-Id: I510ad2bbebb30d055a9d9b99e72d6955e27c73d7