ui: Support fragments
Fragment links (e.g. example.com#foo) are really useful to link to a
specific item in a large page (e.g. a particular flag in the flags
page or widget in the widget page).

Since we use a routeing scheme that uses the original fragment to
encode the page, query string etc we have to implement this ourselves
if we want it.
- This implements fragment links
- Updates widgets_page to use them

Change-Id: I0d4ee42e5562642945875b1e6b06419a77455314