tp: support source replacement in SqlSource
This CL introduces support for rewriting SQL in SqlSource. This is
a requirement for adding support for macros where we will rewrite SQL
statements when expanding the macros but we still want to preserve the
context from the statement that the user wrote.

By modelling this as "recursive rewrites", we can keep track of nested
macro expansions and, when an error happens, print an error message
which takes into account the expansions.

Bug: 290185551
Change-Id: I28da15ffc47f99a6e32ca609d3bcf549968553ca