Neater "recent commands" list using tags.
Currently, recently used commands are repeated in a separate section
at the top of the command palette. The sections are separated by a line
and a header which doesn't look the best!

This CL changes the way this works by simply sorting recently used
commands to the top and adding a "recently used" tag. The sections have
been removed, which looks a lot better IMHO.

Change-Id: Iefdba8a602746adaa85a840f53efb7ba45a418e1