[ui]: Make crticial path slices clickable
We previously checked that the debug slice meets the following
requirements:
1. Has and id and utid
2. The ts and dur of the thread_state from a query with thread_state
id matches the debug slice ts and dur.

The second condition fails with critical path debug slices because the
ts and dur are modified to avoid overlaps with bounding spans in the
critical path.

To fix, we weakened the second constraint by checking for an explicit
table argument in the debug slice.

While at it, did the same for the debug slices from the slice table.
This would be of use for the flattened slices that get their original
ts and dur modified as well.

Test: Manual
Change-Id: I2a8048fc54738668072bbb50f5e03e3d3f5a49e5