message_handle: Implement FinalizationListener
This is going to be used to execute arbitrary code after message
finalization.

The commit is basically a revert of 1aa58ae23fd7("Protozero: remove
message finalization listener").

It's going to be used to update chunks headers when finalizing messages
to allow scraping the last packet from each trace writer.

Bug: 162206162
Change-Id: I0bb69f8b91227f441f2deb77066a20b91d00ef70