service: Add buffer.transfer_on_clone
Add a per-buffer flag which causes the buffer to be
moved into the cloned trace (Rather than copied) and
cleared in the main tracing session.
This is to deal with use-cases when a data source
writes hundreds of MB of data, which is needed only
in the cloned buffer (for bugreport) and that we don't
want to stick around to minimize memory impact of long
traces.
See go/clone-for-winscope for more.

Bug: 293429094
Test: perfetto_unittests --gtest_filter=TracingServiceImplTest.TransferOnClone
Change-Id: I61db4fc49281c0656defadf18878653c81acac54