Increase scheduling niceness of traced/traced_probes
We've recently seen cases where traced & traced_probes don't
get enough cpu bandwidth and cause data drops (b/297046915).
Increase the scheduling priority to mitigate that. We are still
limited to little cores only via ServiceCapacityLow.
When the system is running out of CPU banwidth, it's sensible that
the tracing system that is in charge to tell what's taking CPU
has more priority to actually record the data.
I chose -10 by observing the stackrank of priorities on a recent
Pixel. This puts us above processes and most HALs, below kernel
therads, and right on par with sysui (which, however, runs on big
cores)

Change-Id: Ibd09112a47a302e9631b4e5fabf7e8bcb588e260