Add android_boot metric
For now his computes the total time certain processes spend in
uinterruptible sleep as well as the total duration of all scheduling
slices owned by the given process.  Right now we only care about
uninterruptible sleep, but the proto is designed so that we can easily
append more fields for different scheduling states.  We pick the name
'android_boot' to make it clear that this proto can be used for future
boot related metrics.

Test: <trace_processor_shell> --run-metrics android_boot <trace>
Test: tools/diff_test_trace_processor.py <trace_processor_shell> --name-filter='.*android_boot*'
Bug: 300681738

Change-Id: I5ded2c78cdaab32cbca94bd1b9c1ef9149a32ea5