binder_tracker: Support binder_command and binder_return
The binder_transaction and binder_transaction_received are not enough to
tell what's happening in case of failed transactions. The result is that
we often have unterminated slices.

This commit teaches the binder_tracker to parse binder_command and
binder_return. When these new events are added (in addition to the
existing binder_transaction and binder_transaction_received), the
tracker is able to detect when a transaction fails and correctly
terminate the slice.

Implementation details in the comments.

Bug: 295124679
Change-Id: I1305c019adc6494d7b0756085c5e61162feb568d