[ui] Fix missing tracks in shared permalinks.
When a permalink is loaded, the state is updated before plugins are
initialised. This results in track vnodes & components getting created
before plugins have a chance to register tracks.

For scrolling tracks, tracks are only attempted to be resolved once:
when the track component is constructed. When no track is present at
construct time, the track latches into a state where it shows "No such
track", regardless of whether a plugin registers the correct track later
on.

An elegant fix would be to always register the plugins before attempting
to render any track components, but there didn't seem like an easy way
to do this.

This CL patches the issue by attempting to resolve/create tracks every
DOM render cycle in TrackPanel::view(), until the panel is resolved.
Thus, tracks will appear as soon as they've been registered.

Change-Id: I18f5b8d54fa3bfcf8aab9580ee8a71dd2e76c7f6