[ui] Fix details panel restore button.
Since aosp/2594127, the height of the details panel no longer includes
the height of the drag handle. Before, when the panel was closed it's
`height === DRAG_HANDLE_HEIGHT_PX`, but now `height === 0`. This broke
the logic used to work out whether the panel is closed and thus what to
do when the hide/restore button is pressed.

This CL fixes the test to disregard the drag handle height, and also
tidies up a few other places where the drag handle height was no longer
required.

Bug: 302139005
Change-Id: Id581f38e4843dc566e71022b31e25095a64ad2c6