[ui] Port cpu_freq tracks.
CpuFreqTrackController does some async stuff in it's onSetup function.
Previously this function was being called fire-n-forget, which meant
that onBoundsChange could be called before onSetup was complete. Thus,
this CL adds a more strict lifecycle to tracks where onCreate() is
guaranteed to complete before any of the other methods are called.

Change-Id: I5998fea769741c3218b28817f60cbfb279c3381c