[stdlib]: Add io_wait to critical_path
It's important to distinguish kernel lock contention vs IO, hence
added io_wait information to the critical_path.

Also fixed the blocked_function on the 'self' thread. Previously,
the |critical_path_blocked_function| was used instead of the |self_function|.
This was only correct when the self_state critical_path was resolved.
In IRQ cases on the self thread, the blocked_function is not the same
as the critical_path_blocked_function.

Test: tools/diff_test_trace_processor.py out/android/trace_processor_shell --name-filter '.*thread_executing_span.*'
Change-Id: I136ba316a9622c3ba64020e1dcbf0a72a077d93e