traced: make CloneReadOnly() copy only used portion of the buffer
Before this change CloneReadOnly() would memcpy the whole buffer
when cloning. This has two side-effects:
1. Causes a read-only page-fault on the original buffer. This is
   typically not a huge deal because both Linux and Window will
   map all that memory to the zero-page until a RW fault happens.
2. The cloned buffer will be filled with zeros up to the declared
   size. This can cause memory spikes.

The memory spikes wouldn't be a huge deal normally, because we
expect the buffer to be eventually fully used.
However, recently we introduced edge use cases where we declare
extremely large (~500MB) buffers and expect to use them only in
the case of a bugreport (But not fully). In this case the memory
spike is undesirable.

Bug: 307542353
Test: perfetto_unittests --gtest_filter=TraceBufferTest.*
Change-Id: Ie53ee2b854cf2da488d115623dbdcf959fbacade