docs: Plumb new reference page for the UI plugins API
This is just to set up all the build rules and the plumbing, for now
don't to anything clever and just dump the whole ui/src/public/index.ts
file (e.g. most of the core interfaces with the nice comments) into a
new docs page.

In future we'll do something cleverer and autogenerate a nice reference
page using one of the available tools for TypeScript.

Change-Id: Id06a0c473a32e6b5177077d7f14f82629b79a223